Factor things differently so that the slack details appear everywhere they can
