+++ app/models/behaviors/events/SlackEvent.scala
@@ -2,12 +2,18 @@ package models.behaviors.events
 
 import akka.actor.ActorSystem
 import models.accounts.slack.botprofile.SlackBotProfile
+import play.api.libs.json.{JsArray, JsBoolean, JsObject, JsString}
+import play.api.libs.ws.WSClient
 import services.DataService
+import slack.api.{ApiError, SlackApiClient}
+import slack.models.Channel
 
 import scala.concurrent.ExecutionContext.Implicits.global
+import scala.concurrent.Future
 
 trait SlackEvent {
   val user: String
+  val channel: String
   val profile: SlackBotProfile
   def clientFor(dataService: DataService) = dataService.slackBotProfiles.clientFor(profile)
   def eventualMaybeDMChannel(dataService: DataService)(implicit actorSystem: ActorSystem) = {
@@ -25,6 +31,42 @@ trait SlackEvent {
       ""
     } else {
       s"<@$user>: "
+    }
+  }
+
+  private def maybeChannelInfoFor(client: SlackApiClient)(implicit actorSystem: ActorSystem): Future[Option[Channel]] = {
+    client.getChannelInfo(channel).map(Some(_)).recover {
+      case e: ApiError => None
+    }
+  }
+
+  def detailsFor(ws: WSClient, dataService: DataService)(implicit actorSystem: ActorSystem): Future[JsObject] = {
+    val client = clientFor(dataService)
+    for {
+      user <- client.getUserInfo(user)
+      maybeChannel <- maybeChannelInfoFor(client)
+    } yield {
+      val profileData = user.profile.map { profile =>
+        Seq(
+          profile.first_name.map(v => "firstName" -> JsString(v)),
+          profile.last_name.map(v => "lastName" -> JsString(v)),
+          profile.real_name.map(v => "realName" -> JsString(v))
+        ).flatten
+      }.getOrElse(Seq())
+      val channelMembers = maybeChannel.flatMap { channel =>
+        channel.members.map(_.filterNot(_ == profile.userId))
+      }.getOrElse(Seq())
+      JsObject(
+        Seq(
+          "name" -> JsString(user.name),
+          "profile" -> JsObject(profileData),
+          "isPrimaryOwner" -> JsBoolean(user.is_primary_owner.getOrElse(false)),
+          "isOwner" -> JsBoolean(user.is_owner.getOrElse(false)),
+          "isRestricted" -> JsBoolean(user.is_restricted.getOrElse(false)),
+          "isUltraRestricted" -> JsBoolean(user.is_ultra_restricted.getOrElse(false)),
+          "channelMembers" -> JsArray(channelMembers.map(JsString.apply))
+        )
+      )
     }
   }
